% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{add_to_cache}
\alias{add_to_cache}
\title{Add an entry to the cache}
\usage{
add_to_cache(cache, key, value)
}
\arguments{
\item{cache}{The cache object (list) to add to}

\item{key}{The key to store the value under}

\item{value}{The value to store}
}
\value{
The updated cache object
}
\description{
This function adds or updates a key-value pair in the cache.
}
