% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sibling_repo.R
\name{.abort_repo_not_found}
\alias{.abort_repo_not_found}
\title{Build and throw error for repository not found}
\usage{
.abort_repo_not_found(
  parent_dir,
  repo_name,
  alternative_names,
  current_repo_name
)
}
\description{
Build and throw error for repository not found
}
\keyword{internal}
