% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{.all_rows_identical}
\alias{.all_rows_identical}
\title{Check if all rows in a dataframe are identical}
\usage{
.all_rows_identical(rows_data)
}
\description{
Check if all rows in a dataframe are identical
}
\keyword{internal}
