% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.apply_custom_function}
\alias{.apply_custom_function}
\title{Apply custom grouping function}
\usage{
.apply_custom_function(column_data, custom_fn)
}
\description{
Apply custom grouping function
}
\keyword{internal}
