% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log10.R
\name{.apply_log10_transform}
\alias{.apply_log10_transform}
\title{Apply log10 transformation to variables}
\usage{
.apply_log10_transform(dataf, vars, quiet)
}
\description{
Apply log10 transformation to variables
}
\keyword{internal}
