% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.bin_by_cutpoints}
\alias{.bin_by_cutpoints}
\title{Bin continuous variable using cutpoints}
\usage{
.bin_by_cutpoints(column_data, cutpoints, labels = NULL)
}
\description{
Bin continuous variable using cutpoints
}
\keyword{internal}
