% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.build_groups_from_pattern}
\alias{.build_groups_from_pattern}
\title{Build groups from pattern matching}
\usage{
.build_groups_from_pattern(dataf, pattern_extract)
}
\description{
Build groups from pattern matching
}
\keyword{internal}
