% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_by_group.R
\name{.build_predictors}
\alias{.build_predictors}
\title{Build predictor formula string}
\usage{
.build_predictors(
  base_predictors,
  outcome,
  group,
  outcome_covariates,
  group_covariates,
  model_type,
  interaction_terms,
  is_group_specific
)
}
\description{
Build predictor formula string
}
\keyword{internal}
