% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{.categorize_duplicates}
\alias{.categorize_duplicates}
\title{Categorize duplicates as identical or conflicting}
\usage{
.categorize_duplicates(dataf, id_values, unique_ids)
}
\description{
Categorize duplicates as identical or conflicting
}
\keyword{internal}
