% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.coalesce_groups}
\alias{.coalesce_groups}
\title{Coalesce groups into new columns}
\usage{
.coalesce_groups(dataf, groups, prefix, overwrite, quiet)
}
\description{
Coalesce groups into new columns
}
\keyword{internal}
