% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.create_grouped_variable}
\alias{.create_grouped_variable}
\title{Create grouped variable based on type}
\usage{
.create_grouped_variable(column_data, group)
}
\description{
Create grouped variable based on type
}
\keyword{internal}
