% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.create_groups_from_pattern}
\alias{.create_groups_from_pattern}
\title{Create groups from pattern matching}
\usage{
.create_groups_from_pattern(dataf, pattern_extract)
}
\description{
Create groups from pattern matching
}
\keyword{internal}
