% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_by_group.R
\name{.extract_all_covariates}
\alias{.extract_all_covariates}
\title{Extract all covariate names for validation}
\usage{
.extract_all_covariates(outcome_covariates, group_covariates)
}
\description{
Extract all covariate names for validation
}
\keyword{internal}
