% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_group_comparisons.R
\name{.extract_pairwise_matrix}
\alias{.extract_pairwise_matrix}
\title{Extract p-values from pairwise.t.test matrix into named list}
\usage{
.extract_pairwise_matrix(pvalue_matrix)
}
\description{
Extract p-values from pairwise.t.test matrix into named list
}
\keyword{internal}
