% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_factors.R
\name{.find_cols_for_conversion}
\alias{.find_cols_for_conversion}
\title{Find columns for factor conversion (case-insensitive)}
\usage{
.find_cols_for_conversion(dataf, patterns, exclude)
}
\description{
Find columns for factor conversion (case-insensitive)
}
\keyword{internal}
