% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log10.R
\name{.find_invalid_values}
\alias{.find_invalid_values}
\title{Find variables with invalid values for log transformation}
\usage{
.find_invalid_values(dataf, vars)
}
\description{
Find variables with invalid values for log transformation
}
\keyword{internal}
