% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_by_group.R
\name{.fit_single_model}
\alias{.fit_single_model}
\title{Helper function to build a single model as a tibble}
\usage{
.fit_single_model(
  outcome,
  group,
  predictors,
  data,
  group_col,
  model_type,
  stop_on_error = FALSE
)
}
\description{
Helper function to build a single model as a tibble
}
\keyword{internal}
