% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_impute_data.R
\name{.format_imputation_results}
\alias{.format_imputation_results}
\title{Format imputation results into return list}
\usage{
.format_imputation_results(dataf, imputation_result, helper_vars)
}
\description{
Format imputation results into return list
}
\keyword{internal}
