% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_group_comparisons.R
\name{.format_pvalue_columns}
\alias{.format_pvalue_columns}
\title{Format p-value columns}
\usage{
.format_pvalue_columns(results, pair_names, format, digits)
}
\description{
Format p-value columns
}
\keyword{internal}
