% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_group_comparisons.R
\name{.format_pvalues}
\alias{.format_pvalues}
\title{Format a vector of p-values}
\usage{
.format_pvalues(pvals, format, digits)
}
\description{
Format a vector of p-values
}
\keyword{internal}
