% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.generate_cutpoint_labels}
\alias{.generate_cutpoint_labels}
\title{Generate default labels for cutpoints}
\usage{
.generate_cutpoint_labels(cutpoints, breaks)
}
\description{
Generate default labels for cutpoints
}
\keyword{internal}
