% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_impute_data.R
\name{.get_complete_cases}
\alias{.get_complete_cases}
\title{Get complete cases for error estimation}
\usage{
.get_complete_cases(imputation_data, verbose)
}
\description{
Get complete cases for error estimation
}
\keyword{internal}
