% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sibling_repo.R
\name{.get_parent_directory}
\alias{.get_parent_directory}
\title{Navigate to parent directory from project root}
\usage{
.get_parent_directory(parent_levels)
}
\description{
Navigate to parent directory from project root
}
\keyword{internal}
