% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols_by_pattern.R
\name{.handle_no_matches}
\alias{.handle_no_matches}
\title{Handle case when no columns match}
\usage{
.handle_no_matches(dataf, patterns, exclude, warn_no_match)
}
\description{
Handle case when no columns match
}
\keyword{internal}
