% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_by_group.R
\name{.is_group_specific_structure}
\alias{.is_group_specific_structure}
\title{Check if outcome_covariates uses group-specific structure}
\usage{
.is_group_specific_structure(outcome_covariates, groups)
}
\description{
Check if outcome_covariates uses group-specific structure
}
\keyword{internal}
