% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{.keep_most_complete}
\alias{.keep_most_complete}
\title{Keep most complete row (fewest NAs) for each ID}
\usage{
.keep_most_complete(dataf, id_values, unique_ids, id_col_name)
}
\description{
Keep most complete row (fewest NAs) for each ID
}
\keyword{internal}
