% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_impute_data.R
\name{.prepare_imputation_data}
\alias{.prepare_imputation_data}
\title{Prepare data for imputation}
\usage{
.prepare_imputation_data(dataf, vars_to_impute, helper_vars)
}
\description{
Prepare data for imputation
}
\keyword{internal}
