% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.recode_categorical}
\alias{.recode_categorical}
\title{Recode categorical variable using value mapping}
\usage{
.recode_categorical(column_data, value_map)
}
\description{
Recode categorical variable using value mapping
}
\keyword{internal}
