% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.register_if_enabled}
\alias{.register_if_enabled}
\title{Register processing if enabled}
\usage{
.register_if_enabled(groups, check_processed)
}
\description{
Register processing if enabled
}
\keyword{internal}
