% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{.remove_duplicates}
\alias{.remove_duplicates}
\title{Remove duplicates based on keep strategy}
\usage{
.remove_duplicates(dataf, id_values, dupe_info, keep, id_col)
}
\description{
Remove duplicates based on keep strategy
}
\keyword{internal}
