% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_group_comparisons.R
\name{.run_categorical_comparisons}
\alias{.run_categorical_comparisons}
\title{Run chi-squared/Fisher tests for categorical variables}
\usage{
.run_categorical_comparisons(
  data,
  categorical_vars,
  group_var,
  pairs,
  pair_names,
  p_adjust_method
)
}
\description{
Run chi-squared/Fisher tests for categorical variables
}
\keyword{internal}
