% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_group_comparisons.R
\name{.run_numeric_comparisons}
\alias{.run_numeric_comparisons}
\title{Run pairwise.t.test for numeric variables}
\usage{
.run_numeric_comparisons(
  data,
  numeric_vars,
  group_var,
  pair_names,
  p_adjust_method
)
}
\description{
Run pairwise.t.test for numeric variables
}
\keyword{internal}
