% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sibling_repo.R
\name{.search_for_repo}
\alias{.search_for_repo}
\title{Search for repository by name and alternatives}
\usage{
.search_for_repo(parent_dir, repo_name, alternative_names)
}
\description{
Search for repository by name and alternatives
}
\keyword{internal}
