% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.validate_and_return_manual_groups}
\alias{.validate_and_return_manual_groups}
\title{Validate and return manual groups}
\usage{
.validate_and_return_manual_groups(dataf, var_groups)
}
\description{
Validate and return manual groups
}
\keyword{internal}
