% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.validate_binning_params}
\alias{.validate_binning_params}
\title{Validate bin_and_categorize_variables parameters}
\usage{
.validate_binning_params(dataf, groups)
}
\description{
Validate bin_and_categorize_variables parameters
}
\keyword{internal}
