% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.validate_coalesce_params}
\alias{.validate_coalesce_params}
\title{Validate coalesce_variables parameters}
\usage{
.validate_coalesce_params(
  dataf,
  pattern_extract,
  var_groups,
  prefix,
  overwrite,
  quiet
)
}
\description{
Validate coalesce_variables parameters
}
\keyword{internal}
