% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.validate_group_spec}
\alias{.validate_group_spec}
\title{Validate individual group specification}
\usage{
.validate_group_spec(group, dataf)
}
\description{
Validate individual group specification
}
\keyword{internal}
