% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_and_categorize_variables.R
\name{.validate_label_count}
\alias{.validate_label_count}
\title{Validate label count matches number of groups}
\usage{
.validate_label_count(labels, breaks)
}
\description{
Validate label count matches number of groups
}
\keyword{internal}
