% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_by_group.R
\name{.validate_model_inputs}
\alias{.validate_model_inputs}
\title{Validate basic model inputs}
\usage{
.validate_model_inputs(
  data,
  outcomes,
  base_predictors,
  groups,
  model_type,
  interaction_terms,
  verbose
)
}
\description{
Validate basic model inputs
}
\keyword{internal}
