% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols_by_pattern.R
\name{.validate_pattern_params}
\alias{.validate_pattern_params}
\title{Validate select_cols_by_pattern parameters}
\usage{
.validate_pattern_params(
  dataf,
  patterns,
  exclude,
  ignore_case,
  warn_no_match,
  invert
)
}
\description{
Validate select_cols_by_pattern parameters
}
\keyword{internal}
