% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log10.R
\name{.warn_if_log_named}
\alias{.warn_if_log_named}
\title{Warn if variables already have log10 naming convention}
\usage{
.warn_if_log_named(vars)
}
\description{
Warn if variables already have log10 naming convention
}
\keyword{internal}
