% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.warn_if_overwriting}
\alias{.warn_if_overwriting}
\title{Warn if column will be overwritten}
\usage{
.warn_if_overwriting(new_name, dataf)
}
\description{
Warn if column will be overwritten
}
\keyword{internal}
