% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_factors.R
\name{.warn_no_matches}
\alias{.warn_no_matches}
\title{Warn when no columns match patterns}
\usage{
.warn_no_matches(patterns, exclude)
}
\description{
Warn when no columns match patterns
}
\keyword{internal}
