% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_stats}
\alias{get_cache_stats}
\title{Get cache statistics}
\usage{
get_cache_stats(cache)
}
\arguments{
\item{cache}{List containing the cache data}
}
\value{
A list with statistics about the cache
}
\description{
Provides statistics about the cache contents.
}
