% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_from_cache}
\alias{get_from_cache}
\title{Retrieve a value from cache}
\usage{
get_from_cache(cache, key, default = NULL)
}
\arguments{
\item{cache}{The cache object (list) to retrieve from}

\item{key}{The key to look up in the cache}

\item{default}{The default value to return if key is not found}
}
\value{
The cached value or the default if not found
}
\description{
This function safely retrieves a value from a cache object
by key, returning a default value if the key doesn't exist.
}
