% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_items_to_fetch}
\alias{get_items_to_fetch}
\title{Find items that need to be fetched (not in cache)}
\usage{
get_items_to_fetch(items, cache)
}
\arguments{
\item{items}{Vector of items to check against the cache}

\item{cache}{The cache object (list) to check}
}
\value{
Vector of items not found in the cache
}
\description{
This function compares a list of items against the cache
to determine which ones need to be fetched. This is useful
for batch processing and API calls.
}
