% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{get_processed_vars_env}
\alias{get_processed_vars_env}
\title{Get environment for tracking processed variables}
\usage{
get_processed_vars_env()
}
\value{
An environment object for storing processing status
}
\description{
Get environment for tracking processed variables
}
\keyword{internal}
