% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{initialize_cache}
\alias{initialize_cache}
\title{Initialize a cache}
\usage{
initialize_cache(cache_name, cache_dir = "cache", extension = ".rds")
}
\arguments{
\item{cache_name}{Character string for the base name of the cache file}

\item{cache_dir}{Character string for the directory where cache files are stored}

\item{extension}{Character string for the file extension}
}
\value{
A list containing the cache data
}
\description{
Loads an existing cache file or creates a new one if it doesn't exist.
}
