% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{is_nonempty_character}
\alias{is_nonempty_character}
\title{Assert a non-empty character vector}
\usage{
is_nonempty_character()
}
\value{
An assertion object consumed by \code{\link[=validate_args]{validate_args()}}.
}
\description{
Use inside \code{\link[=validate_args]{validate_args()}} to check that an argument is a character
vector of length at least 1.
}
\examples{
\dontrun{
validate_args(vars = is_nonempty_character())
}
}
