% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{is_nonempty_list}
\alias{is_nonempty_list}
\title{Assert a non-empty list}
\usage{
is_nonempty_list()
}
\value{
An assertion object consumed by \code{\link[=validate_args]{validate_args()}}.
}
\description{
Use inside \code{\link[=validate_args]{validate_args()}} to check that an argument is a list of
length at least 1.
}
\examples{
\dontrun{
validate_args(groups = is_nonempty_list())
}
}
