% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{is_one_of}
\alias{is_one_of}
\title{Assert one of a fixed set of values}
\usage{
is_one_of(choices)
}
\arguments{
\item{choices}{A vector of allowed values.}
}
\value{
An assertion object consumed by \code{\link[=validate_args]{validate_args()}}.
}
\description{
Use inside \code{\link[=validate_args]{validate_args()}} to check that an argument matches one of
a set of allowed values.
}
\examples{
\dontrun{
validate_args(method = is_one_of(c("bonferroni", "holm", "BH")))
}
}
