% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{is_processed}
\alias{is_processed}
\title{Check if variables have been processed}
\usage{
is_processed(func_name, vars, error_if_exists = FALSE)
}
\arguments{
\item{func_name}{String. Name of the processing function}

\item{vars}{Character vector. Names of variables being processed}

\item{error_if_exists}{Logical. If TRUE, throws an error if already processed}
}
\value{
Logical. TRUE if variables have been processed, FALSE otherwise
}
\description{
Checks whether the specified variables have already been processed
by a given function.
}
