% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_cache}
\alias{save_cache}
\title{Save cache to disk}
\usage{
save_cache(
  cache,
  cache_name,
  cache_dir = "cache",
  extension = ".rds",
  force = FALSE,
  periodic = TRUE,
  save_frequency = 100,
  counter = NULL
)
}
\arguments{
\item{cache}{The cache object to save}

\item{cache_name}{Name of the cache}

\item{cache_dir}{Directory where cache files are stored, defaults to "cache"}

\item{extension}{File extension to use, defaults to ".rds"}

\item{force}{Logical, force saving regardless of counter}

\item{periodic}{Logical, enable periodic saving based on counter}

\item{save_frequency}{How often to save when periodic is TRUE, default is 100}

\item{counter}{Current iteration counter for periodic saving}
}
\value{
Logical, TRUE if cache was saved, FALSE otherwise
}
\description{
This function saves a cache to disk, either immediately or
based on a periodic counter. This allows for periodic saving during
long-running operations to prevent data loss.
}
