% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{add_to_cache}
\alias{add_to_cache}
\title{Add entry to cache}
\usage{
add_to_cache(cache, key, value)
}
\arguments{
\item{cache}{Cache object}

\item{key}{Key for the entry}

\item{value}{Value to cache}
}
\value{
Updated cache object
}
\description{
Add entry to cache
}
