% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clean_cache}
\alias{clean_cache}
\title{Clean expired entries from cache}
\usage{
clean_cache(cache, max_age_days = 30)
}
\arguments{
\item{cache}{Cache object}

\item{max_age_days}{Maximum age in days}
}
\value{
Cleaned cache object
}
\description{
Clean expired entries from cache
}
