% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_factors.R
\name{.find_matching_columns}
\alias{.find_matching_columns}
\title{Find columns matching patterns with optional exclusions}
\usage{
.find_matching_columns(dataf, patterns, exclude)
}
\description{
Find columns matching patterns with optional exclusions
}
\keyword{internal}
