% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{.load_csv_files}
\alias{.load_csv_files}
\title{Load CSV files into data frames}
\usage{
.load_csv_files(
  file_paths,
  missing_vals,
  clean_col_names,
  verbose,
  use_readr,
  col_types
)
}
\description{
Load CSV files into data frames
}
\keyword{CSV}
\keyword{Load}
\keyword{data}
\keyword{files}
\keyword{frames}
\keyword{internal}
\keyword{into}
