% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_variables.R
\name{.should_skip_processing}
\alias{.should_skip_processing}
\title{Check if processing should be skipped}
\usage{
.should_skip_processing(groups, dataf, force, check_processed)
}
\description{
Check if processing should be skipped
}
\keyword{internal}
