% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_path}
\alias{get_cache_path}
\title{Create a standardized cache file path}
\usage{
get_cache_path(
  cache_name,
  cache_dir = "cache",
  extension = ".rds",
  create_dir = TRUE,
  validate_name = TRUE
)
}
\arguments{
\item{cache_name}{Character string for the cache name (will be sanitized)}

\item{cache_dir}{Directory to store cache files, defaults to "cache"}

\item{extension}{File extension to use, defaults to ".rds"}

\item{create_dir}{Logical, should the directory be created if missing?}

\item{validate_name}{Logical, should cache_name be validated/sanitized?}
}
\value{
A full path to the cache file
}
\description{
Create a standardized cache file path
}
