% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{initialize_cache}
\alias{initialize_cache}
\title{Initialize a cache}
\usage{
initialize_cache(cache_name, cache_dir = "cache")
}
\arguments{
\item{cache_name}{Name for the cache}

\item{cache_dir}{Directory for cache files (default: "cache")}
}
\value{
Cache object (list)
}
\description{
Initialize a cache
}
