% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{initialize_cache}
\alias{initialize_cache}
\title{Initialize or load a cache}
\usage{
initialize_cache(cache_name, cache_dir = "cache", extension = ".rds")
}
\arguments{
\item{cache_name}{Name of the cache}

\item{cache_dir}{Directory where cache files are stored, defaults to "cache"}

\item{extension}{File extension to use, defaults to ".rds"}
}
\value{
A list containing the cache data
}
\description{
This function creates a new cache or loads an existing one from disk. If the cache doesn't exist, it will return an empty list. If it does
exist it will load the cache from the file.
}
