% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_cache_expired}
\alias{is_cache_expired}
\title{Check if cache is expired}
\usage{
is_cache_expired(cache, max_age_days = 30)
}
\arguments{
\item{cache}{Cache object}

\item{max_age_days}{Maximum age in days}
}
\value{
Logical
}
\description{
Check if cache is expired
}
