% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates_if_exists}
\alias{remove_duplicates_if_exists}
\title{Remove duplicate rows based on ID}
\usage{
remove_duplicates_if_exists(
  dataf,
  id_col,
  keep = c("first", "last", "none"),
  quiet = FALSE
)
}
\arguments{
\item{dataf}{A data frame to check for duplicates}

\item{id_col}{Column name or expression specifying the ID to check for duplicates}

\item{keep}{One of "first" (default), "last", or "none". Determines which occurrence
to keep when duplicates are found.}

\item{quiet}{Logical; if TRUE, suppresses messages about duplicate IDs}
}
\value{
A data frame with duplicates removed according to the 'keep' parameter
}
\description{
This function identifies and removes duplicate rows based on a specified ID column.
It provides information about which IDs have duplicates and whether the duplicate
rows are identical or have conflicting values.
}
\details{
The keep argumet will match full or partially,
ie "la" = "last", "al" = "all", "fi", "fir" = "first"
}
\examples{
testDF <- data.frame(
id_num = c(1, 1, 2, 2, 3, 4, 5, 5, 6),
value = c(100, 100, 200, 250, 300, 400, 500, 500, 600),
group = c("A", "A", "B", "B", "C", "D", "E", "F", "G")
)
remove_duplicates_if_exists(testDF, "id_num", "first")
remove_duplicates_if_exists(testDF, "id_num", "fir")
}
