% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_cache}
\alias{save_cache}
\title{Save cache to disk}
\usage{
save_cache(
  cache,
  cache_name,
  cache_dir = "cache",
  extension = ".rds",
  force = FALSE,
  periodic = TRUE,
  save_frequency = 100,
  counter = NULL,
  compress = FALSE
)
}
\arguments{
\item{cache}{List containing the cache data}

\item{cache_name}{Character string for the base name of the cache file}

\item{cache_dir}{Character string for the directory where cache files are stored}

\item{extension}{Character string for the file extension}

\item{force}{Logical indicating whether to force saving regardless of other conditions}

\item{periodic}{Logical indicating whether to use periodic saving based on counter}

\item{save_frequency}{Integer indicating how often to save when using periodic saving}

\item{counter}{Integer indicating the current iteration count}

\item{compress}{Logical indicating whether to use compression when saving, default = FALSE}
}
\value{
Logical indicating whether the cache was saved
}
\description{
Saves a cache object to disk, with options for periodic saving and compression.
}
